Support installing combined core-edge cluster

+++ common.sh
@@ -37,27 +37,33 @@ function installCluster() {
       echo "Directory already exists: $dir"
       echo "Stopping previous installation"
       $dir/bin/neo4j stop
-    else
-      echo "Installing $dir"
-      mkdir -p $dir
-      cp -a $NEO4J_TEMPLATE/* $dir/
-      $dir/bin/neo4j-admin set-initial-password $NEO4J_PASSWORD
+      echo "Deleting previous installation"
+      rm -Rf $dir
     fi
+    echo "Installing $dir"
+    mkdir -p $dir
+    cp -a $NEO4J_TEMPLATE/* $dir/
+    $dir/bin/neo4j-admin set-initial-password $NEO4J_PASSWORD
   done
 }
 
 function deleteCluster() {
   prefix=$1
-  typeset -i i END # Let's be explicit
-  let END=$2
-  for ((i=1;i<=END;++i)); do
-    dir="${prefix}_${i}"
-    if [ -d "$dir" ] ; then
-      rm -Rf $dir
-    else
-      echo "Directory does not exist: $dir"
-    fi
-  done
+  if [ -d "${prefix}_1" ] ; then
+    for dir in ${prefix}_? ; do
+      if [ -d "$dir" ] ; then
+        echo "Found installation at $dir - stopping and deleting it"
+        if [ -x "$dir/bin/neo4j" ] ; then
+          $dir/bin/neo4j stop
+        fi
+        rm -Rf $dir
+      else
+        echo "Not a directory: $dir"
+      fi
+    done
+  else
+    echo "No ${prefix} installations found"
+  fi
 }
 
 function printConfigFromPatch() {
@@ -69,7 +75,23 @@ function printConfigFromPatch() {
   done
 }
 
+function makeDiscoveryMembersConfig() {
+  prefix="core"
+  members=""
+  typeset -i i END # Let's be explicit
+  let END=$NUMBER_CORES
+  for ((i=1;i<=END;++i)); do
+    port="500$i"
+    if [ -n "$members" ] ; then
+      members="${members},"
+    fi
+    members="${members}localhost:$port"
+  done
+  echo $members
+}
+
 function configureCluster() {
+  members=$(makeDiscoveryMembersConfig)
   prefix=$1
   ce_mode=$(echo "$prefix" | tr '[:lower:]' '[:upper:]')
   typeset -i i END # Let's be explicit
@@ -80,11 +102,12 @@ function configureCluster() {
       echo "Configuring $dir"
       mkdir -p $dir/conf
       cp -a $NEO4J_TEMPLATE/conf/neo4j.conf $dir/conf/
-      conf_patch="neo4j.conf.${prefix}_${instance}.patch"
+      conf_patch="neo4j.conf.${prefix}_${i}.patch"
       echo -e "\tMaking patch for ${prefix}_${i} in $conf_patch"
-      cat neo4j.conf.patch \
+      cat neo4j.conf.${prefix}.patch \
         | sed -e "s/^\+\(.*\)1$/\+\1$i/" \
-        | sed -e "s/^\+dbms.mode=CORE$/\+dbms.mode=$ce_mode/" \
+        | sed -e "s/^\+core_edge.expected_core_cluster_size=3$/\+core_edge.expected_core_cluster_size=$NUMBER_CORES/" \
+        | sed -e "s/^\+core_edge.initial_discovery_members=.*$/\+core_edge.initial_discovery_members=$members/" \
         > $conf_patch
       if [ -f "$conf_patch" ] ; then
         echo -e "\tPatching $dir/conf/neo4j.conf"
@@ -154,7 +177,7 @@ function clearCluster() {
     dir="${prefix}_${i}"
     if [ -d "$dir" ] ; then
       echo "Clearing logs and graph.db for $dir"
-      rm -f $dir/logs/* $dir/data/databases/graph.db
+      rm -Rf $dir/logs/* $dir/data/databases/graph.db
     else
       echo "Directory does not exist: $dir"
     fi
@@ -166,7 +189,11 @@ function waitForCluster() {
   typeset -i i END # Let's be explicit
   let END=$2
   for ((i=1;i<=END;++i)); do
-    http_port="748${i}"
+    if [ "$prefix" = "core" ] ; then
+      http_port="748${i}"
+    else
+      http_port="749${i}"
+    fi
     end="$((SECONDS+20))"
     echo -en "\tWaiting for response at port $http_port "
     rc=0
